[refactor] Rename "untravel" module
