+++ packages/prosemirror-lwdita/src/tests/document.spec.ts
@@ -18,7 +18,7 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 import ChaiPromised from 'chai-as-promised';
 import { use, expect, assert } from 'chai';
 import { document, NODES, _test_private_document } from '../document';
-import { unTravel } from '../untravelNodes';
+import { unTravel } from '../untravel-document';
 import {
   JDITA_OBJECT,
   TRANSFORMED_JDITA_OBJECT,
